#447 Separate migration CLI tools in a separate module.
