+++ pom.xml
@@ -118,6 +118,7 @@
         <module>dao</module>
         <module>menas</module>
         <module>migrations</module>
+        <module>migrations-cli</module>
         <module>external/kafka</module>
         <module>external/fixedWidth</module>
         <module>testutils</module>
